+++ src/modules/llms/store-llms.ts
@@ -158,6 +158,7 @@ function updateSelectedIds(allLlms: DLLM[], chatLlmId: DLLMId | null, fastLlmId:
 export function findLLMOrThrow(llmId: DLLMId): DLLM {
   const llm = useModelsStore.getState().llms.find(llm => llm.id === llmId);
   if (!llm) throw new Error(`LLM ${llmId} not found`);
+  if (!llm._source) throw new Error(`LLM ${llmId} has no source`);
   return llm;
 }
 
@@ -166,7 +167,6 @@ export function findOpenAILlmRefOrThrow(llmId: DLLMId): string {
   if (!openAIModelRef) throw new Error(`LLM ${llmId} has no OpenAI LLM`);
   return openAIModelRef;
 }
-
 
 
 export function useChatLLM() {
